Update README.md
The documentation previously linked to API_EXAMPLE.go, which does not exist and resulted in a 404 error. This has been updated to correctly point to API_EXAMPLE_1.go in the src directory.

Additional broken or outdated links have been observed throughout the repository and will be listed separately for batch correction.